samplernn-scripts/requirements.txt
librosa
essentia
pyacoustid
numpy
pysoundfile
